+++ lib/core/app_error_codes.dart
@@ -1,0 +1,7 @@
+class AppErrorCodes {
+  static int noError = 0;
+  static int generalError = 1;
+  static int commandCannotExecute = 126;
+  static int commandNotFound = 127;
+  static int fatalError = 128;
+}