+++ lib/core/commands/gen_refs_command.dart
@@ -5,6 +5,7 @@ import 'package:args/command_runner.dart';
 import '../../core/app_constants.dart';
 import '../../entities/config.dart';
 import '../../usecases/create_file_references.dart';
+import '../app_error_codes.dart';
 
 class GenRefsCommand extends Command {
   final Config config;
@@ -23,12 +24,26 @@ class GenRefsCommand extends Command {
         abbr: AppConstants.argShortPackageName,
         defaultsTo: null,
         mandatory: false);
+
+    argParser.addOption(AppConstants.argLongExcludeSuffix,
+        abbr: AppConstants.argShortExcludeSuffix,
+        defaultsTo: AppConstants.excludeSuffixDefaultValue,
+        mandatory: false);
+
+    argParser.addMultiOption(AppConstants.argLongExcludeFiles,
+        splitCommas: true, help: AppConstants.argLongExcludeFilesHelp);
+
+    argParser.addMultiOption(AppConstants.argLongExcludeContents,
+        splitCommas: true, help: AppConstants.argLongExcludeContentsHelp);
+
+    argParser.addOption(AppConstants.argLongExcludeContentsPath,
+        help: AppConstants.argLongExcludeContentsPathHelp, mandatory: false);
   }
 
   /// excute run createFileReferences
   @override
   void run() async {
     await createFileReferences();
-    exit(0);
+    exit(AppErrorCodes.noError);
   }
 }